Moved `smart_open` function into utils
